Backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document Processing
pypdf==3.17.1
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1

# AI and Vector Search
openai==1.3.0
faiss-cpu==1.7.4
numpy==1.26.2
pandas==2.1.3

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.4.2

# Testing
pytest==7.4.3
httpx==0.25.1 